requirement.txt
jiwer==2.5.1
ffmpeg==1.4
numpy==1.23.4
datasets==2.7.1
zmq==0.0.0
torchaudio==0.12.1
tqdm==4.64.1
transformer==4.24.0
numpy==1.23.4